Avances y cargar malla de pixeles copiada

+++ WhittedRaytracer/WhittedRaytracer/Camara.cpp
@@ -2,7 +2,7 @@
 
 Camara::Camara() {}
 
-Camara::Camara(vec3 posicion, vec3 normal, vec3 centro, double lejos, double cerca, double fov_vertical, double fov_horizontal) {
+Camara::Camara(vec3 posicion, vec3 normal, vec3 centro, float lejos, float cerca, float fov_vertical, float fov_horizontal) {
 	this->centro = centro;
 	this->posicion = posicion;
 	this->cerca = cerca;
